requirements.txt
amadeus==5.0.0
Django==5.0
django_formtools==2.5.1
djangorestframework==3.14.0
langchain_core==0.1.45
langchain_openai==0.1.3
pytest==7.4.3
Requests==2.31.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3